pyproject.toml
[project]
name = "lightning-agents"
version = "0.1.0"
description = "Factory-of-Factories pattern for dynamic Claude Agent instantiation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
lightning = "lightning_agents.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]
presentation = [
    "python-pptx>=1.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
